requirements.txt
Django==2.2.4
djangorestframework==3.10.2
jieba==0.39
